// projects/vbr-image-edit/src/lib/classes/key.enum.ts
export const enum Key {
  Backspace = 8,
  Tab = 9,
  Enter = 13,
  Shift = 16,
  Control = 17,
  Alt = 18,
  CapsLock = 20,
  Escape = 27,
  Space = 32,
  PageUp = 33,
  PageDown = 34,
  End = 35,
  Home = 36,
  ArrowLeft = 37,
  ArrowUp = 38,
  ArrowRight = 39,
  ArrowDown = 40,

  Semicolon = 186,
  Colon = 186,
  EqualsSign = 187,
  Plus = 187,
  Comma = 188,
  LessThanSign = 188,
  Minus = 189,
  Underscore = 189,
  Period = 190,
  GreaterThanSign = 190,
  ForwardSlash = 191,
  QuestionMark = 191,
  Backtick = 192,
  Tilde = 192,
  OpeningSquareBracket = 219,
  OpeningCurlyBrace = 219,
  Backslash = 220,
  Pipe = 220,
  ClosingSquareBracket = 221,
  ClosingCurlyBrace = 221,
  SingleQuote = 222,
  DoubleQuote = 222,

  Pause = 19,
  PrintScreen = 44,
  Insert = 45,
  Delete = 46,
  Num0 = 48,
  Num1 = 49,
  Num2 = 50,
  Num3 = 51,
  Num4 = 52,
  Num5 = 53,
  Num6 = 54,
  Num7 = 55,
  Num8 = 56,
  Num9 = 57,
  A = 65,
  B = 66,
  C = 67,
  D = 68,
  E = 69,
  F = 70,
  G = 71,
  H = 72,
  I = 73,
  J = 74,
  K = 75,
  L = 76,
  M = 77,
  N = 78,
  O = 79,
  P = 80,
  Q = 81,
  R = 82,
  S = 83,
  T = 84,
  U = 85,
  V = 86,
  W = 87,
  X = 88,
  Y = 89,
  Z = 90,
  MetaLeft = 91,
  MetaRight = 92,
  ContextMenu = 93,
  Numpad0 = 96,
  Numpad1 = 97,
  Numpad2 = 98,
  Numpad3 = 99,
  Numpad4 = 100,
  Numpad5 = 101,
  Numpad6 = 102,
  Numpad7 = 103,
  Numpad8 = 104,
  Numpad9 = 105,
  NumpadMultiply = 106,
  NumpadAdd = 107,
  NumpadSubtract = 109,
  NumpadDecimal = 110,
  NumpadDivide = 111,
  F1 = 112,
  F2 = 113,
  F3 = 114,
  F4 = 115,
  F5 = 116,
  F6 = 117,
  F7 = 118,
  F8 = 119,
  F9 = 120,
  F10 = 121,
  F11 = 122,
  F12 = 123,
  NumLock = 144,
  ScrollLock = 145
}